archive/project/pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Atlas, Adaptive Thought & Learning Autonomous System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pygame>=2.5",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "gymnasium>=0.29",
  "torch>=2.1",
  "stable-baselines3>=2.3",
  "sb3-contrib>=2.3",
  "sqlalchemy>=2.0",
  "networkx>=3.2",
  "openai>=1.12",
  "PyYAML>=6.0",
]

[project.scripts]
atlas = "src.main:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
